Add: MapSprite.getTileAtPosition

+++ QuickTiGame2d/QuickTiGame2d/QuickTiGame2dMapSprite.h
@@ -74,7 +74,7 @@
 -(BOOL)updateTileCount;
 
 -(NSArray*)getTiles;
--(QuickTiGame2dMapTile*)getTileAtPosition:(NSInteger)sx sy:(NSInteger)sy;
+-(QuickTiGame2dMapTile*)getTileAtPosition:(float)sx sy:(float)sy;
 -(QuickTiGame2dMapTile*)getTile:(NSInteger)index;
 -(void)setTile:(NSInteger)index gid:(NSInteger)gid;
 -(void)setTile:(NSInteger)index tile:(QuickTiGame2dMapTile*)tile;
